驱动部分代码除蓝牙串口，其余部分代码已经移植完毕，这是平衡小车的基础驱动代码；

+++ Drivers/Input/rotary_encoder.h
@@ -38,5 +38,5 @@
 ///** @} */
 
 void Encoder_Init(void);
-int16_t Encoder_Get(void);
+int8_t Encoder_Get(void);
 #endif
